+++ generators/scaffold/dashboard/dashboard.go
@@ -2,6 +2,7 @@ package dashboard
 
 import (
 	"embed"
+	"os/exec"
 	"text/template"
 
 	"github.com/swiftcarrot/dashi/generators/scaffold"
@@ -9,7 +10,7 @@ import (
 	"github.com/swiftcarrot/genny/gogen"
 )
 
-//go:embed templates
+//go:embed templates/$pages$ templates/$components$
 var templates embed.FS
 
 func New(opts *scaffold.Options) (*genny.Generator, error) {
@@ -23,12 +24,14 @@ func New(opts *scaffold.Options) (*genny.Generator, error) {
 
 	name := opts.Name.Pascalize().Pluralize().ToLower().String()
 
-	g.Transformer(genny.Replace("-pages-", "packages/dashboard/src/pages/"+name))
-	g.Transformer(genny.Replace("-components-", "packages/components/src/"+name))
+	g.Transformer(genny.Replace("$pages$", "packages/dashboard/src/pages/"+name))
+	g.Transformer(genny.Replace("$components$", "packages/components/src/"+name))
 
 	if err := g.Templates(&templates); err != nil {
 		return g, err
 	}
+
+	g.Command(exec.Command("yarn", "fmt"))
 
 	return g, nil
 }